Add route and link to referat container. Made dialogmoteside container.
